+++ components/tags/list.tsx
@@ -3,7 +3,7 @@ interface Props {
   title: string;
   description: string;
 }
-const TagList = ({ children, title, description }: Props) => {
+const List = ({ children, title, description }: Props) => {
   return (
     <div>
       <div>{title}</div>
@@ -13,4 +13,4 @@ const TagList = ({ children, title, description }: Props) => {
   );
 };
 
-export default TagList;
+export default List;
