+++ pages/tags/index.tsx
@@ -1,6 +1,6 @@
 import Layout from "@/components/layout";
-import TagItem from "@/components/tags/tag-item";
-import TagList from "@/components/tags/tag-list";
+import Item from "@/components/tags/item";
+import List from "@/components/tags/list";
 import useSWR from "swr";
 import { Tag } from "@prisma/client";
 import Link from "next/link";
@@ -12,11 +12,11 @@ const Tags = () => {
     <Layout>
       <div className="p-4">
         <Link href="/tags/create">태그 추가</Link>
-        <TagList title="책 태그" description="장르, 특징, 목적 등">
+        <List title="책 태그" description="장르, 특징, 목적 등">
           {data?.tags.map((tag: Tag) => (
-            <TagItem key={tag.id} tag={tag} />
+            <Item key={tag.id} tag={tag} />
           ))}
-        </TagList>
+        </List>
       </div>
     </Layout>
   );
